Boo YAML Mustache/Ini4J configuration feature
The feature is described in a bit of documentation added to the README.md.

We use Ini4J to read a ~/.oneops/config file that is in the same form as the
AWS config file. The values from the "default" profile are loaded into a Map<String,String>
and interpolated into the specified Boo YAML template using the Java implementation
of Mustache by Sam Pullara.

We can easily add features like reading public keys from a Git repo and making it
available as a variable, or easily allow the use of alternative profiles in the
config. Anything that can be processed as text can be made available in
the Mustache context and injected into the template.